requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
Django==4.2.4
django-environ==0.10.0
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.7
PyJWT==2.8.0
pytz==2023.3
redis==4.6.0
sqlparse==0.4.4
typing_extensions==4.7.1

xlwt~=1.3.0